moving callbacks, related to custom markdown tags into a separate folder
