+++ src/callbacks/html/methods/advanced/image.js
@@ -1,4 +1,4 @@
-import { commonReplace } from '../../../domain/replace-wrapper3.0/commonReplace';
+import { commonReplace } from '../../../../domain/replace-wrapper3.0/commonReplace';
 
 // TODO remove unused `text` argument
 export function _image(text, alt, srcWithTooltip) {
